Refactor security features into specializations of authorized feature
- disconnect registry from tagged directory in IAASRegistryDecorator
