+++ basyx.components/basyx.components.docker/basyx.components.AASServer/src/main/resources/security.properties
@@ -7,7 +7,7 @@
 # ######################
 # authorization.strategy=SimpleRbac
 authorization.strategy=GrantedAuthority
-authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider=org.eclipse.basyx.components.security.authorization.KeycloakJwtBearerTokenAuthenticationConfigurationProvider
+authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider=org.eclipse.basyx.components.security.authorization.internal.KeycloakJwtBearerTokenAuthenticationConfigurationProvider
 authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider.keycloak.serverUrl=http://localhost:9005
 authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider.keycloak.realm=basyx-demo
 authorization.strategy.simpleRbac.rulesFilePath=/rbac_rules.json
