+++ basyx.components/basyx.components.docker/basyx.components.registry/src/main/java/org/eclipse/basyx/components/registry/authorization/internal/IAuthorizersProvider.java
@@ -22,23 +22,26 @@
  *
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
-package org.eclipse.basyx.components.security.authorization;
+package org.eclipse.basyx.components.registry.authorization.internal;
 
 import org.eclipse.basyx.components.configuration.BaSyxSecurityConfiguration;
-import org.eclipse.basyx.vab.protocol.http.server.JwtBearerTokenAuthenticationConfiguration;
+import org.eclipse.basyx.components.registry.authorization.internal.Authorizers;
 
 /**
- * Provider for {@link JwtBearerTokenAuthenticationConfiguration}, which will be passed into the BaSyx server context to be used as a security filter and set up the security context for incoming requests. Uses the aas server configuration.
+ * Provides the authorizers for the different common to be authorized BaSyx
+ * objects using the registry server configuration.
  *
  * @author wege
  */
-public interface IJwtBearerTokenAuthenticationConfigurationProvider {
+public interface IAuthorizersProvider<SubjectInformationType> {
 	/**
-	 * Provides the {@link JwtBearerTokenAuthenticationConfiguration} that can be passed to the BaSyx server context to install a security filter and validate and set up the security context from access tokens included in incoming
-	 * requests.
+	 * Provides the authorizers for the different common to be authorized BaSyx
+	 * objects using the AAS server configuration.
 	 *
-	 * @param securityConfig the configuration of the aas server which should have information on how to determine the {@link JwtBearerTokenAuthenticationConfiguration}.
-	 * @return the {@link JwtBearerTokenAuthenticationConfiguration} object
+	 * @param securityConfig
+	 *            the aas server configuration that holds information about how to
+	 *            determine the authorizers.
+	 * @return the different authorizers bundled in an {@link Authorizers}
 	 */
-	public JwtBearerTokenAuthenticationConfiguration get(BaSyxSecurityConfiguration securityConfig);
+	public Authorizers<SubjectInformationType> get(BaSyxSecurityConfiguration securityConfig);
 }
