+++ basyx.components/basyx.components.docker/basyx.components.registry/src/main/resources/security.properties
@@ -7,7 +7,7 @@
 # ######################
 # authorization.strategy=SimpleRbac
 authorization.strategy=GrantedAuthority
-authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider=org.eclipse.basyx.components.security.authorization.KeycloakJwtBearerTokenAuthenticationConfigurationProvider
+authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider=org.eclipse.basyx.components.security.authorization.internal.KeycloakJwtBearerTokenAuthenticationConfigurationProvider
 authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider.keycloak.serverUrl=http://127.0.0.1:9005/auth
 authorization.strategy.jwtBearerTokenAuthenticationConfigurationProvider.keycloak.realm=basyx-demo
 authorization.strategy.simpleRbac.rulesFilePath=/rbac_rules.json
