+++ basyx.components/basyx.components.lib/src/main/java/org/eclipse/basyx/components/security/authorization/internal/IJwtBearerTokenAuthenticationConfigurationProvider.java
@@ -22,26 +22,23 @@
  *
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
-package org.eclipse.basyx.components.security.authorization;
+package org.eclipse.basyx.components.security.authorization.internal;
 
 import org.eclipse.basyx.components.configuration.BaSyxSecurityConfiguration;
-import org.eclipse.basyx.extensions.shared.authorization.internal.KeycloakService;
 import org.eclipse.basyx.vab.protocol.http.server.JwtBearerTokenAuthenticationConfiguration;
 
 /**
- * Implementation of the {@link IJwtBearerTokenAuthenticationConfigurationProvider} interface.
- * <p>
- * Provides the {@link JwtBearerTokenAuthenticationConfiguration} that can be used to configure the BaSyx server to accept tokens from a specific Keycloak server and realm and set up the security context on an incoming request.
+ * Provider for {@link JwtBearerTokenAuthenticationConfiguration}, which will be passed into the BaSyx server context to be used as a security filter and set up the security context for incoming requests. Uses the aas server configuration.
  *
  * @author wege
  */
-public class KeycloakJwtBearerTokenAuthenticationConfigurationProvider implements IJwtBearerTokenAuthenticationConfigurationProvider {
-	@Override public JwtBearerTokenAuthenticationConfiguration get(BaSyxSecurityConfiguration securityConfig) {
-		final String serverUrl = securityConfig.getAuthorizationStrategyJwtBearerTokenAuthenticationConfigurationProviderKeycloakServerUrl();
-		final String realm = securityConfig.getAuthorizationStrategyJwtBearerTokenAuthenticationConfigurationProviderKeycloakRealm();
-
-		final KeycloakService keycloakService = new KeycloakService(serverUrl, realm);
-
-		return keycloakService.createJwtBearerTokenAuthenticationConfiguration();
-	}
+public interface IJwtBearerTokenAuthenticationConfigurationProvider {
+	/**
+	 * Provides the {@link JwtBearerTokenAuthenticationConfiguration} that can be passed to the BaSyx server context to install a security filter and validate and set up the security context from access tokens included in incoming
+	 * requests.
+	 *
+	 * @param securityConfig the configuration of the aas server which should have information on how to determine the {@link JwtBearerTokenAuthenticationConfiguration}.
+	 * @return the {@link JwtBearerTokenAuthenticationConfiguration} object
+	 */
+	public JwtBearerTokenAuthenticationConfiguration get(BaSyxSecurityConfiguration securityConfig);
 }
